requirements.txt
numpy==1.26.4
opencv-python==4.8.1.78
pillow==10.2.0
pytesseract
rich
